Check for updates using the state hash
The state hash is stored in the deployment under the state_hash
annotation. When an update event is triggered a new state hash
is created from some of the Nextcloud object attributes and
checked against the state hash on the annotations.

We have an object update from the user if the state hash generated
from the object is different from the one in the annotations. And
the deploy function is called.